Partage de la recordsheet au format CSV

+++ app/src/main/java/fr/villot/pricetracker/activities/MainActivity.java
@@ -1,46 +1,31 @@
 package fr.villot.pricetracker.activities;
 
-import android.Manifest;
-import android.app.Activity;
-import android.content.Intent;
-import android.content.pm.PackageManager;
 import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
 
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.ActionBar;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
-import androidx.core.app.ActivityCompat;
-import androidx.core.content.ContextCompat;
-import androidx.core.view.GravityCompat;
-import androidx.drawerlayout.widget.DrawerLayout;
 import androidx.fragment.app.Fragment;
 import androidx.viewpager.widget.ViewPager;
 
 import com.google.android.material.tabs.TabLayout;
 
-import java.io.File;
-import java.util.logging.Logger;
+import java.util.Objects;
 
+import fr.villot.pricetracker.R;
 import fr.villot.pricetracker.adapters.PageAdapter;
 import fr.villot.pricetracker.fragments.ProductsFragment;
 import fr.villot.pricetracker.fragments.RecordSheetsFragment;
 import fr.villot.pricetracker.fragments.StoresFragment;
-import fr.villot.pricetracker.R;
 
 public class MainActivity extends AppCompatActivity {
-    private TabLayout tabLayout;
-    private ViewPager viewPager;
-    private PageAdapter pageAdapter;
-    private Toolbar toolbar;
-    DrawerLayout drawerLayout;
+    //    DrawerLayout drawerLayout;
     private boolean isSelectionModeActive = false;
     private Fragment currentFragment;
 
-    private static final Logger logger = Logger.getLogger(MainActivity.class.getName());
+//    private static final Logger logger = Logger.getLogger(MainActivity.class.getName());
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -48,23 +33,23 @@ public class MainActivity extends AppCompatActivity {
         setContentView(R.layout.activity_main);
 
         // Recuperation des vues
-        tabLayout = findViewById(R.id.tabLayout);
-        viewPager = findViewById(R.id.viewPager);
-        toolbar = findViewById(R.id.toolbar);
-        drawerLayout = findViewById(R.id.drawerLayout);
+        TabLayout tabLayout = findViewById(R.id.tabLayout);
+        ViewPager viewPager = findViewById(R.id.viewPager);
+        Toolbar toolbar = findViewById(R.id.toolbar);
+//        drawerLayout = findViewById(R.id.drawerLayout);
 
         // Toolbar
         setSupportActionBar(toolbar);
 
-        // Ajout du bouton hamburger
-        ActionBar actionBar = getSupportActionBar();
-        if (actionBar != null) {
-            actionBar.setDisplayHomeAsUpEnabled(true);
-            actionBar.setHomeAsUpIndicator(R.drawable.ic_menu);
-        }
+//        // Ajout du bouton hamburger
+//        ActionBar actionBar = getSupportActionBar();
+//        if (actionBar != null) {
+//            actionBar.setDisplayHomeAsUpEnabled(true);
+//            actionBar.setHomeAsUpIndicator(R.drawable.ic_menu);
+//        }
 
         // Gestion de la navigation par onglets
-        pageAdapter = new PageAdapter(getSupportFragmentManager());
+        PageAdapter pageAdapter = new PageAdapter(getSupportFragmentManager());
         viewPager.setAdapter(pageAdapter);
         tabLayout.setupWithViewPager(viewPager);
         tabLayout.setTabMode(TabLayout.MODE_FIXED);
@@ -91,49 +76,52 @@ public class MainActivity extends AppCompatActivity {
     public boolean onCreateOptionsMenu(Menu menu) {
         MenuInflater inflater = getMenuInflater();
 //        inflater.inflate(R.menu.main_menu, menu);
-        if (isSelectionModeActive)
+        if (isSelectionModeActive) {
             inflater.inflate(R.menu.toolbar_selection_menu, menu);
+            if (currentFragment instanceof RecordSheetsFragment) {
+                MenuItem itemShare = menu.getItem(0);
+                itemShare.setVisible(true);
+            }
+        }
 
         return true;
     }
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
-        // Handle menu item clicks
-        switch (item.getItemId()) {
-            case android.R.id.home:
-                if (isSelectionModeActive) {
-                    setSelectionMode(currentFragment, false);
-                } else {
-                    // Ouverture du menu déroulant hamburger
-                    drawerLayout.openDrawer(GravityCompat.START);
-                }
-                return true;
-            case R.id.action_settings:
-                return true;
-            case R.id.action_share:
-                if (currentFragment instanceof RecordSheetsFragment) {
-                    RecordSheetsFragment recordSheetsFragment = (RecordSheetsFragment) currentFragment;
-                    recordSheetsFragment.shareRecordSheet();
-                }
-                return true;
-            case R.id.action_delete:
-                if (currentFragment instanceof ProductsFragment) {
-                    ProductsFragment productsFragment = (ProductsFragment) currentFragment;
-                    productsFragment.deleteSelectedItems();
-                }
-                else if (currentFragment instanceof StoresFragment) {
-                    StoresFragment storesFragment = (StoresFragment) currentFragment;
-                    storesFragment.deleteSelectedItems();
-                }
-                else if (currentFragment instanceof RecordSheetsFragment) {
-                    RecordSheetsFragment recordSheetsFragment = (RecordSheetsFragment) currentFragment;
-                    recordSheetsFragment.deleteSelectedItems();
-                }
-                return true;
-            default:
-                return super.onOptionsItemSelected(item);
+        int itemId = item.getItemId();
+
+        if (itemId == android.R.id.home) {
+            if (isSelectionModeActive) {
+                setSelectionMode(currentFragment, false);
+            }
+//          else {
+//                // Ouverture du menu déroulant hamburger
+//                drawerLayout.openDrawer(GravityCompat.START);
+//            }
+            return true;
+        } else if (itemId == R.id.action_share) {
+
+            if (currentFragment instanceof RecordSheetsFragment) {
+                RecordSheetsFragment recordSheetsFragment = (RecordSheetsFragment) currentFragment;
+                recordSheetsFragment.shareRecordSheet();
+            }
+            return true;
+        } else if (itemId == R.id.action_delete) {
+            if (currentFragment instanceof ProductsFragment) {
+                ProductsFragment productsFragment = (ProductsFragment) currentFragment;
+                productsFragment.deleteSelectedItems();
+            } else if (currentFragment instanceof StoresFragment) {
+                StoresFragment storesFragment = (StoresFragment) currentFragment;
+                storesFragment.deleteSelectedItems();
+            } else if (currentFragment instanceof RecordSheetsFragment) {
+                RecordSheetsFragment recordSheetsFragment = (RecordSheetsFragment) currentFragment;
+                recordSheetsFragment.deleteSelectedItems();
+            }
+            return true;
         }
+
+        return super.onOptionsItemSelected(item);
     }
 
     // Gestion du mode de sélection
@@ -146,7 +134,8 @@ public class MainActivity extends AppCompatActivity {
         if (this.isSelectionModeActive != isSelectionModeActive) {
             this.isSelectionModeActive = isSelectionModeActive;
             if (!isSelectionModeActive) {
-                getSupportActionBar().setHomeAsUpIndicator(R.drawable.ic_menu);
+                Objects.requireNonNull(getSupportActionBar()).setDisplayHomeAsUpEnabled(false);
+//                getSupportActionBar().setHomeAsUpIndicator(R.drawable.ic_back);
                 getSupportActionBar().setTitle(R.string.app_name);
 
                 if (currentFragment instanceof ProductsFragment) {
@@ -163,6 +152,7 @@ public class MainActivity extends AppCompatActivity {
                 }
 
             } else {
+                Objects.requireNonNull(getSupportActionBar()).setDisplayHomeAsUpEnabled(true);
                 getSupportActionBar().setHomeAsUpIndicator(R.drawable.ic_cancel);
             }
             invalidateOptionsMenu(); // Rafraichissement du menu de la toolbar
